{42} player fragment needs to update (#45)
* {40} Reworked the PlayerFragment´s Layout.

* {40} PlayerService can now change it´s volume on a scale between 0 and 1. Values passed to the PlayerService are adjusted with the function f(x) = 10^(1-1/(x^0.75)).
The PlayerFragment now has Buttons to change the volume of the PlayerService.

* {42} PlayerFragment´s progressbar now works. It should also now always display the correct information about the currently playing song. The text displayed on the PlayerFragment is also not editable anymore.

* {42} The PlayerService now actually gets stopped after a couple of seconds idling.

+++ app/src/main/java/de/teamawesome/awesomeplayer/playerservice/PlayerListenerConvenienceImplementation.java
@@ -30,7 +30,7 @@ public abstract class PlayerListenerConvenienceImplementation implements IPlayba
     }
 
     @Override
-    public void volumeChanged(int newVolume) {
+    public void volumeChanged(float newVolume) {
 
     }
 }
